moving all patches work to core module
